stattracker_api/api/serializers.py
from rest_framework import serializers
from .models import Activity

class ActivitySerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Activity
        fields = ('id', 'activity_name','activity_date', 'activity_count')
